Revert "Revert "change naming style and update functionality""
This reverts commit 4d93ecdf63467a4311f894483d91d64de53574ea.
