+++ integration/config/region_service_exclusion.yaml
@@ -1,6 +1,5 @@
 regions:
   af-south-1:
-    exclude_resources:
     - ServerlessRepo
     - Cognito
     - KMS
@@ -10,16 +9,13 @@ regions:
     - GatewayResponses
     - HttpApi
   ap-east-1:
-    exclude_resources:
     - Cognito
     - IoT
     - ServerlessRepo
     - HttpApi
   ap-northeast-2:
-    exclude_resources:
     - HttpApi
   ap-northeast-3:
-    exclude_resources:
     - Cognito
     - IoT
     - ServerlessRepo
@@ -27,18 +23,14 @@ regions:
     - CodeDeploy
     - HttpApi
   ap-south-1:
-    exclude_resources:
     - HttpApi
   ap-southeast-1:
-    exclude_resources:
     - HttpApi
   ca-central-1:
-    exclude_resources:
     - Cognito
     - IoT
     - HttpApi
   cn-north-1:
-    exclude_resources:
     - ServerlessRepo
     - Cognito
     - KMS
@@ -48,7 +40,6 @@ regions:
     - GatewayResponses
     - HttpApi
   cn-northwest-1:
-    exclude_resources:
     - ServerlessRepo
     - Cognito
     - KMS
@@ -58,14 +49,12 @@ regions:
     - GatewayResponses
     - HttpApi
   eu-north-1:
-    exclude_resources:
     - ServerlessRepo
     - Cognito
     - IoT
     - HttpApi
     - Layers
   eu-south-1:
-    exclude_resources:
     - ServerlessRepo
     - Cognito
     - KMS
@@ -75,29 +64,23 @@ regions:
     - GatewayResponses
     - HttpApi
   eu-west-2:
-    exclude_resources:
     - HttpApi
   eu-west-3:
-    exclude_resources:
     - Cognito
     - IoT
     - XRay
     - HttpApi
   me-south-1:
-    exclude_resources:
     - ServerlessRepo
     - Cognito
     - IoT
     - HttpApi
   sa-east-1:
-    exclude_resources:
     - IoT
     - Cognito
     - HttpApi
   us-east-2:
-    exclude_resources:
     - HttpApi
   us-west-1:
-    exclude_resources:
     - Cognito
     - IoT
