+++ integration/single/test_basic_http_api.py
@@ -1,7 +1,7 @@
 from unittest.case import skipIf
 
 from integration.helpers.base_test import BaseTest
-from integration.helpers.resource import should_exclude_test_in_region
+from integration.helpers.resource import current_region_does_not_support
 
 
 class TestBasicHttpApi(BaseTest):
@@ -9,7 +9,7 @@ class TestBasicHttpApi(BaseTest):
     Basic AWS::Serverless::HttpApi tests
     """
 
-    @skipIf(should_exclude_test_in_region("HttpApi"), "HttpApi is not supported in this testing region")
+    @skipIf(current_region_does_not_support(["HttpApi"]), "HttpApi is not supported in this testing region")
     def test_basic_http_api(self):
         """
         Creates a HTTP API
