+++ integration/single/test_basic_layer_version.py
@@ -1,10 +1,10 @@
 from unittest.case import skipIf
 
 from integration.helpers.base_test import BaseTest
-from integration.helpers.resource import should_exclude_test_in_region
+from integration.helpers.resource import current_region_does_not_support
 
 
-@skipIf(should_exclude_test_in_region("Layers"), "Layers is not supported in this testing region")
+@skipIf(current_region_does_not_support(["Layers"]), "Layers is not supported in this testing region")
 class TestBasicLayerVersion(BaseTest):
     """
     Basic AWS::Lambda::LayerVersion tests
